recognition/requirements.txt
boto3==1.34.122
pillow==10.4.0
google-generativeai==0.7.2
